Add files via upload
The journey to becoming a successful data analyst requires dedication, continuous learning, and the right skill set. Following this step-by-step roadmap will provide you with a comprehensive guide to developing the necessary skills and knowledge. Embrace the data-driven era and equip yourself with the tools and techniques to excel as a data analyst in 2025 and beyond.